The function get_face_embedding is moved to utils
